sample-application/requirements.txt
streamlit
hvac
boto3
python-dotenv
awscli